+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Use tree-sitter Node.js bindings instead of web (WASM) bindings.
 
+### Removed
+- Support for Node.js 17 removed - see [tree-sitter/tree-sitter#1503](https://github.com/tree-sitter/tree-sitter/issues/1503)
+
 ## [0.5.1] - 2022-02-11
 ### Fixed
 - Fix server immediately crashing when starting [#30](https://github.com/cucumber/language-server/pull/30)