+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.6.0] - 2022-04-26
 ### Changed
 - Use tree-sitter Node.js bindings instead of web (WASM) bindings.
 - Updated to `@cucumber/language-service 0.13.0`
@@ -64,7 +65,8 @@ code is. The server uses this for auto complete and detecting defined steps.
 ### Added
 - First release
 
-[Unreleased]: https://github.com/cucumber/language-server/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/cucumber/language-server/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/cucumber/language-server/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/cucumber/language-server/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/cucumber/language-server/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/cucumber/language-server/compare/v0.3.3...v0.4.0